i2c driver / i2c Lua module
* i2c_setup, speed argument in hertzs for coherence.
* i2c.attach, speed argument in hertzs for coherence.
* updated i2c.lua example (eeprom 4c256d) implemented ack polling
* i2c.write now accepts "n" arguments. Each argument can be a byte, or a string

+++ components/lua_rtos/drivers/sensor.h
@@ -142,7 +142,7 @@ typedef struct {
 
 		struct {
 			uint8_t  id;
-			uint16_t speed;
+			uint32_t speed;
 			int8_t  sda;
 			int8_t  scl;
 			uint16_t address;
